requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pillow==10.1.0
numpy==1.24.3
ultralytics==8.0.228
requests==2.31.0
opencv-python-headless==4.8.1.78
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
